+++ trunk/open-ig/src/hu/openig/model/AIWorld.java
@@ -288,7 +288,7 @@ public class AIWorld {
 	 */
 	public int inventoryCount(ResearchType rt) {
 		Integer i = inventory.get(rt);
-		return i != null ? i.intValue() : 0;
+		return i != null ? i : 0;
 	}
 	/**
 	 * Add a specific amount to the current inventory level.
@@ -297,7 +297,7 @@ public class AIWorld {
 	 */
 	public void addInventoryCount(ResearchType type, int count) {
 		Integer c = inventory.get(type);
-		int ci = c != null ? c.intValue() : 0;
+		int ci = c != null ? c : 0;
 		int ci2 = ci + count;
 		if (ci2 <= 0) {
 			inventory.remove(type);
